+++ module/torgeternity.js
@@ -824,6 +824,10 @@ async function rollSkillMacro(skillName, attributeName, isInteractionAttack, DND
 Hooks.on('renderCombatTracker', (combatTracker) => {
   const hands = game.cards;
   for (const hand of hands) {
+    // This causes the combat tracker to be "close()"d when the hand is deleted,
+    // but is required to update the CombatTracker when a hand is changed so that
+    // the "pooled" cards can be displayed.
+    // torgeternityCards._onDelete will remove the combat tracker from hand.apps
     hand.apps[combatTracker.id] = combatTracker;
   }
 });